+++ config/src/main/java/org/hyperledger/besu/config/StubGenesisConfigOptions.java
@@ -396,16 +396,6 @@ public class StubGenesisConfigOptions implements GenesisConfigOptions, Cloneable
   }
 
   @Override
-  public boolean isQuorum() {
-    return false;
-  }
-
-  @Override
-  public OptionalLong getQip714BlockNumber() {
-    return OptionalLong.empty();
-  }
-
-  @Override
   public PowAlgorithm getPowAlgorithm() {
     return isEthHash() ? PowAlgorithm.ETHASH : PowAlgorithm.UNSUPPORTED;
   }