+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/response/JsonRpcErrorResponse.java
@@ -14,6 +14,7 @@
  */
 package org.hyperledger.besu.ethereum.api.jsonrpc.internal.response;
 
+import java.util.Arrays;
 import java.util.Objects;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
@@ -26,10 +27,16 @@ public class JsonRpcErrorResponse implements JsonRpcResponse {
 
   private final Object id;
   private final JsonRpcError error;
+  @JsonIgnore private final RpcErrorType errorType;
 
   public JsonRpcErrorResponse(final Object id, final JsonRpcError error) {
     this.id = id;
     this.error = error;
+    this.errorType = findErrorType(error.getCode(), error.getMessage());
+  }
+
+  public JsonRpcErrorResponse(final Object id, final RpcErrorType error) {
+    this(id, new JsonRpcError(error));
   }
 
   @JsonGetter("id")
@@ -57,7 +64,7 @@ public class JsonRpcErrorResponse implements JsonRpcResponse {
       return false;
     }
     final JsonRpcErrorResponse that = (JsonRpcErrorResponse) o;
-    return Objects.equals(id, that.id) && error == that.error;
+    return Objects.equals(id, that.id) && Objects.equals(error, that.error);
   }
 
   @Override
@@ -68,5 +75,17 @@ public class JsonRpcErrorResponse implements JsonRpcResponse {
   @Override
   public String toString() {
     return MoreObjects.toStringHelper(this).add("id", id).add("error", error).toString();
+  }
+
+  @JsonIgnore
+  public RpcErrorType getErrorType() {
+    return errorType;
+  }
+
+  private RpcErrorType findErrorType(final int code, final String message) {
+    return Arrays.stream(RpcErrorType.values())
+        .filter(e -> e.getCode() == code && e.getMessage().equals(message))
+        .findFirst()
+        .get();
   }
 }